renaming to init
