+++ cmd/replicate/main.go
@@ -6,8 +6,8 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/replicate/cli/internal/cmd"
-	"github.com/replicate/cli/internal/cmd/clone"
 	"github.com/replicate/cli/internal/cmd/hardware"
+	init_cmd "github.com/replicate/cli/internal/cmd/init"
 	"github.com/replicate/cli/internal/cmd/model"
 	"github.com/replicate/cli/internal/cmd/prediction"
 	"github.com/replicate/cli/internal/cmd/training"
@@ -38,7 +38,7 @@ func init() {
 		prediction.RootCmd,
 		training.RootCmd,
 		hardware.RootCmd,
-		clone.RootCmd,
+		init_cmd.RootCmd,
 	} {
 		rootCmd.AddCommand(cmd)
 		cmd.GroupID = "core"
